Fix multi word variable bug
Thanks FatBoyXPC for this commit
This was creating an issue where `docker compose` was being read
as a single variable with a space, rather than two separate words.
This can be fixed in a few ways, such as putting something like
`setopt shwordsplit` in the user's zshrc, but that requires every
user to have that set. Interestingly enough, eval doesn't seem to
have this same issue, only the backticks and $() do.